buf: make len an uint8_t
The buffer content's maximum length is known to be <= 120 and thus fit
into a single byte. There's no need to use up 8/16 bytes. As a result,
`struct buf` no longer weights nice and round 128 bytes on 64bit
platforms.

+++ status.c
@@ -13,7 +13,7 @@
 
 
 struct buf {
-    size_t len;
+    uint8_t len;
     char data[120];
 };
